+++ src/sll/debug_object.c
@@ -19,6 +19,7 @@ static sll_node_t* _remove_debug_data_internal(sll_node_t* o){
 		case SLL_NODE_TYPE_FLOAT:
 		case SLL_NODE_TYPE_FIELD:
 		case SLL_NODE_TYPE_FUNCTION_ID:
+		case SLL_NODE_TYPE_DECL_COPY:
 			return o+1;
 		case SLL_NODE_TYPE_FUNC:
 		case SLL_NODE_TYPE_INTERNAL_FUNC: