+++ src/sll/util.c
@@ -51,7 +51,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_string_index_t sll_add_string(sll_string_t
 
 
 
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_string_index_t sll_add_string_runtime(sll_string_table_t* st,sll_object_t* v){
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_string_index_t sll_add_string_object(sll_string_table_t* st,sll_object_t* v){
 	SLL_ASSERT(SLL_OBJECT_GET_TYPE(v)==SLL_OBJECT_TYPE_STRING);
 	for (sll_string_index_t i=0;i<st->l;i++){
 		if (sll_string_equal(st->dt+i,&(v->dt.s))){